pyproject.toml
[tool.poetry]
name = "sredoshahmaty"
version = "0.1.0"
description = ""
authors = ["Vadim Evard <v.e.evard@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.27.2"
pytz = "^2024.2"
python-dotenv = "^1.0.1"
pendulum = "^3.0.0"
aiogram = "^3.13.1"
jinja2 = "^3.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
